restructure file for preparation
