+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom/client";
 import "./index.css";
 import reportWebVitals from "./reportWebVitals";
-import { App } from "./App";
+import { App } from "./components/App";
 
 const root = ReactDOM.createRoot(document.getElementById("root")!);
 root.render(
